# openapi-specs/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dating API",
    "description": "An ASP.NET Core Web API for managing Dating items",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Example Contact",
      "url": "https://example.com/contact"
    },
    "license": {
      "name": "Example License",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Members/{id}": {
      "get": {
        "tags": [
          "Members"
        ],
        "operationId": "Members_GetMember",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Member"
                }
              }
            }
          }
        }
      }
    },
    "/api/Members/{id}/photos": {
      "get": {
        "tags": [
          "Members"
        ],
        "operationId": "Members_GetPhotos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Photo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Members/photo": {
      "post": {
        "tags": [
          "Members"
        ],
        "operationId": "Members_AddPhoto",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "ContentType": {
                    "type": "string",
                    "nullable": true
                  },
                  "ContentDisposition": {
                    "type": "string",
                    "nullable": true
                  },
                  "Headers": {
                    "type": "array",
                    "nullable": true,
                    "items": {}
                  },
                  "Length": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "Name": {
                    "type": "string",
                    "nullable": true
                  },
                  "FileName": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Photo"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Member": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "birthDay": {
            "type": "string",
            "format": "date"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastActive": {
            "type": "string",
            "format": "date-time"
          },
          "gender": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "Photo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          },
          "publicId": {
            "type": "string",
            "nullable": true
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "memberId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}